refactor: use long ids for repositories

+++ src/main/java/com/library/library/borrow/BorrowRepository.java
@@ -8,7 +8,7 @@ import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 
-public interface BorrowRepository extends JpaRepository<Borrow, Integer> {
+public interface BorrowRepository extends JpaRepository<Borrow, Long> {
     public Optional<Borrow> findByBookAndMemberAndDelivery(Book book, Member member, boolean delivery);
 
     public List<Borrow> findAllByDeliverAndDateBetween(boolean deliver, LocalDate from, LocalDate to);
